Update to Spring Boot 2 and Spring Cloud Finchley
Move legacy endpoint configuration to separate class.

Fixes gh-102.
