+++ cereebro-snitch/src/main/java/io/cereebro/snitch/actuate/LegacyCereebroEndpointConfiguration.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright © 2017 the original authors (http://cereebro.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.cereebro.snitch.actuate;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.actuate.endpoint.annotation.Endpoint;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@ConditionalOnWebApplication
+@ConditionalOnClass({Endpoint.class, WebMvcConfigurer.class})
+public class LegacyCereebroEndpointConfiguration {
+    
+    @Value("${management.endpoints.web.base-path:/actuator}")
+    private String basePath;
+    
+    @Bean
+    WebMvcConfigurer cereebroSnitchLegacyEndpointConfigurer() {
+        return new WebMvcConfigurer() {
+            @Override
+            public void addViewControllers(ViewControllerRegistry registry) {
+                // To keep backward compatibility with previous servers.
+                // Since Spring Boot 2 by default Cereebro is exposed as /actuator/cereebro
+                // There is no easy way to change default url mapping to /actuator/cereebro/snitch in cereebro-snitch code unless
+                // explicitly changed in service using 'management.endpoints.web.path-mapping.cereebro=/actuator/cereebro/snitch'.
+                registry.addViewController(basePath + CereebroSnitchMvcEndpoint.DEFAULT_PATH)
+                        .setViewName("forward:" + basePath + "/cereebro");
+            }
+        };
+    }
+}
